+++ src/Catalog/Review/ReviewService.php
@@ -40,6 +40,7 @@ final class ReviewService extends AbstractService
     private const COMPANY_ENDPOINT = "catalog/companies/%s/reviews";
     private const PRODUCT_ENDPOINT = "catalog/products/%s/reviews";
     private const AUTHORIZATION_ENDPOINT = "catalog/companies/%s/reviews/authorized";
+    private const AUTHORIZATION_PRODUCT_ENDPOINT = "catalog/products/%s/reviews/authorized";
 
     /**
      * @return Review[]
@@ -112,6 +113,20 @@ final class ReviewService extends AbstractService
                 return false;
             }
 
+            throw $e;
+        }
+
+        return true;
+    }
+
+    public function canUserReviewProduct(string $productId) : bool
+    {
+        try {
+            $this->client->get(sprintf(self::AUTHORIZATION_PRODUCT_ENDPOINT, $productId));
+        } catch (ClientException $e) {
+            if ($e->getCode() === 401) {
+                return false;
+            }
             throw $e;
         }