+++ tests/Review/ReviewServiceTest.php
@@ -134,6 +134,12 @@ final class ReviewServiceTest extends ApiTestCase
         $this->assertFalse($response);
     }
 
+    public function testCannotUserReviewProduct()
+    {
+        $response = $this->reviewService->canUserReviewProduct((string) 5);
+        $this->assertFalse($response);
+    }
+
     private function buildReviewService(): ReviewService
     {
         $client = $this->buildApiClient();